fix: move socket io and udp to `connectivity.js` plugin section
