--- packages/nyerati/functions/index.js
@@ -1,7 +1,0 @@
-const parsecoord = require("./parsecoord");
-const socket = require("./socket");
-
-module.exports = {
-  parsecoord,
-  socket,
-};
